chore: delete types.ts
ButtonPropsOverrides interfaces are in ButtonTheme.ts
