+++ src/components/button/index.ts
@@ -1,2 +1,1 @@
 export * from './ButtonTheme';
-export * from './types';
